chore(atomix): use compatible serializers
