+++ atomix/utils/src/main/java/io/atomix/utils/serializer/serializers/DefaultSerializers.java
@@ -22,7 +22,8 @@ import io.atomix.utils.serializer.Serializer;
 public final class DefaultSerializers {
 
   /** Basic serializer. */
-  public static final Serializer BASIC = Serializer.builder().build();
+  public static final Serializer BASIC =
+      Serializer.builder().withCompatibleSerialization(true).build();
 
   private DefaultSerializers() {}
 }
