+++ upgrade-tests/src/test/java/io/zeebe/test/UpgradeTest.java
@@ -27,6 +27,7 @@ import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import org.agrona.IoUtil;
 import org.assertj.core.util.Files;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.RuleChain;
@@ -169,6 +170,7 @@ public class UpgradeTest {
         });
   }
 
+  @Ignore
   @Test
   public void oldGatewayWithNewBroker() {
     // given
